REFACTOR component structure: Fix file name
