+++ frontend/src/Components/CardSelector.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent } from '@testing-library/preact';
 import { SCALES, VOTE_COFFEE, VOTE_NOTE_VOTED } from '../constants';
-import { getRenderWithWebSocket } from '../test-helpers/renderWithWebSocker';
+import { getRenderWithWebSocket } from '../test-helpers/renderWithWebSocket';
 import { CardSelector } from './CardSelector';
 
 const render = getRenderWithWebSocket(<CardSelector />, {
