+++ frontend/src/Components/RevealButton.test.tsx
@@ -1,4 +1,4 @@
-import { getRenderWithWebSocket } from '../test-helpers/renderWithWebSocker';
+import { getRenderWithWebSocket } from '../test-helpers/renderWithWebSocket';
 import { fireEvent } from '@testing-library/preact';
 import { RevealButton } from './RevealButton';
 import { SCALES } from '../constants';
